Finished Report!
Finished the frame stuff. Fixed a couple typos on the Euler sets. Also!
The Fixed angles that you had used a different convention than the text
book taught us so I removed it from the report :I Comparing them to the
ones on the handout, they are no where near the same.